+++ packages/browser-client/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @montelo/browser-client
 
+## 1.1.75
+
+### Patch Changes
+
+- personal org
+
 ## 1.1.74
 
 ### Patch Changes